planning_app/requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.1
django-admin-generator==2.6.0
django-restframework==0.0.1
django-utils2==3.0.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-utils==3.8.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
